+++ src/main/java/com/backend/labpoint/dto/subject/SubjectDTO.java
@@ -3,5 +3,5 @@ package com.backend.labpoint.dto.subject;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record SubjectDTO(@NotNull Integer id, @NotBlank String name) {
+public record SubjectDTO(@NotNull Long id, @NotBlank String name) {
 }
